[web] Fixed lazy exports (#619)
# Why

Default import was removed from react-native-web. https://github.com/necolas/react-native-web/pull/1277
The use of module.exports and import/export breaks tree-shaking in with babel-preset-expo.

Related: https://github.com/kmagiera/react-native-gesture-handler/pull/617
Fixes: https://github.com/expo/expo/issues/4339

# How

* Moved the shims to `Gestures(.web).js`
* Moved all modules out of main re-export file. (optional)
