+++ State.js
@@ -1,4 +1,4 @@
-export default {
+const State = {
   UNDETERMINED: 0,
   FAILED: 1,
   BEGAN: 2,
@@ -6,3 +6,14 @@ export default {
   ACTIVE: 4,
   END: 5,
 };
+
+State.print = state => {
+  const keys = Object.keys(State);
+  for (let i = 0; i < keys.length; i++) {
+    if (state === State[keys[i]]) {
+      return keys[i];
+    }
+  }
+};
+
+export default State;
