+++ package.json
@@ -24,6 +24,13 @@
     "Directions.web.js",
     "DrawerLayout.js",
     "GestureHandler.js",
+    "Gestures.js",
+    "GestureButtons.js",
+    "GestureComponents.js",
+    "GestureComponents.web.js",
+    "createNativeWrapper.js",
+    "GestureHandlerPropTypes.js",
+    "NativeViewGestureHandler.js",
     "GestureHandlerButton.js",
     "GestureHandlerButton.web.js",
     "gestureHandlerRootHOC.android.js",
